fix routing for admin panel
